requirements.txt
torch>=1.12.0
transformers>=4.21.0
numpy>=1.21.0
psutil>=5.8.0
pyyaml>=6.0
accelerate>=0.20.0
safetensors>=0.3.0
sentencepiece>=0.1.95
tokenizers>=0.13.0 